Move auth to module
