+++ api/lib/user.js
@@ -93,33 +93,6 @@ class User extends Generic {
         return this.deserialize(pgres.rows[0]);
     }
 
-
-    static async is_auth(req) {
-        if (req.user === 'internal') return true;
-
-        if (!req.user || !req.user.access || !req.user.id) {
-            throw new Err(401, null, 'Authentication Required');
-        }
-
-        if (req.user.access === 'disabled') {
-            throw new Err(403, null, 'Account Disabled - Please Contact Us');
-        }
-
-        return true;
-    }
-
-    static async is_admin(req) {
-        if (req.user === 'internal') return true;
-
-        await this.is_auth(req);
-
-        if (req.user.access !== 'admin') {
-            throw new Err(401, null, 'Admin token required');
-        }
-
-        return true;
-    }
-
     async commit(pool) {
         try {
             await pool.query(sql`
@@ -174,7 +147,7 @@ class User extends Generic {
                     ${user.username},
                     ${user.email},
                     ${await bcrypt.hash(user.password, 10)},
-                    'user'
+                    ${user.access || 'user'}
                 ) RETURNING *
             `);
