+++ api/routes/mapbox.js
@@ -1,6 +1,6 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
 
@@ -20,7 +20,7 @@ async function router(schema, config) {
         res: 'res.Mapbox.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             res.json({
                 token: config.Mapbox
