+++ api/routes/meta.js
@@ -1,6 +1,6 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 const Meta = require('../lib/meta');
 const Settings = require('../lib/settings');
 
@@ -21,7 +21,7 @@ async function router(schema, config) {
         res: 'res.ListMeta.json'
     }, async (req, res) => {
         try {
-            await User.is_admin(req);
+            await Auth.is_admin(req);
 
             res.json(await Meta.list(config.pool, req.query));
         } catch (err) {
@@ -44,7 +44,7 @@ async function router(schema, config) {
         res: 'res.Meta.json'
     }, async (req, res) => {
         try {
-            await User.is_admin(req);
+            await Auth.is_admin(req);
 
             const meta = await Settings.generate(config.pool, req.body);
 
@@ -72,7 +72,7 @@ async function router(schema, config) {
         res: 'res.Meta.json'
     }, async (req, res) => {
         try {
-            await User.is_admin(req);
+            await Auth.is_admin(req);
 
             const meta = await Settings.from(config.pool, req.params.key);
             if (!(meta instanceof Meta)) throw new Err(400, null, 'Meta not found');
@@ -105,7 +105,7 @@ async function router(schema, config) {
         res: 'res.Standard.json'
     }, async (req, res) => {
         try {
-            await User.is_admin(req);
+            await Auth.is_admin(req);
 
             const meta = await Meta.from(config.pool, req.params.key);
             await meta.delete(config.pool);
