+++ api/routes/project-access.js
@@ -1,7 +1,7 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
 const ProjectAccess = require('../lib/project/access');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
     /**
@@ -23,7 +23,7 @@ async function router(schema, config) {
         res: 'res.ListProjectAccess.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             res.json(await ProjectAccess.list(config.pool, req.params.pid, req.query));
         } catch (err) {
@@ -51,7 +51,7 @@ async function router(schema, config) {
         res: 'res.ListProjectAccess.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             req.body.pid = req.params.pid;
             const access = await ProjectAccess.generate(config.pool, req.body);
@@ -81,7 +81,7 @@ async function router(schema, config) {
         res: 'res.ProjectAccess.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const access = await ProjectAccess.from(config.pool, req.params.id);
             access.patch(req.body);
