+++ api/routes/project-iteration-export.js
@@ -6,8 +6,7 @@ const s3 = new AWS.S3({ region: process.env.AWS_DEFAULT_REGION });
 const Submission = require('../lib/project/iteration/submission');
 const createCsvStringifier = require('csv-writer').createObjectCsvStringifier;
 const RL = require('readline');
-const User = require('../lib/user');
-
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
     /**
@@ -29,7 +28,7 @@ async function router(schema, config) {
     }, async (req, res) => {
         try {
             req.user = req.token;
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             if (req.query.threshold && req.query.inferences === 'all') throw new Err(400, null, 'Threshold can only set if inferences is not "all"');
