+++ api/routes/project-iteration-stack-queue.js
@@ -6,7 +6,7 @@ const Iteration = require('../lib/project/iteration');
 const Imagery = require('../lib/project/imagery');
 const Submission = require('../lib/project/iteration/submission');
 const CWAlarm = require('../lib/cw-alarm');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
     const alarm = new CWAlarm(config);
@@ -29,7 +29,7 @@ async function router(schema, config) {
         res: 'res.StackQueue.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             const queue = await StackQueue.from(req.params.pid, req.params.iterationid);
@@ -57,7 +57,7 @@ async function router(schema, config) {
         res: 'res.StackQueue.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             await StackQueue.delete(req.params.pid, req.params.iterationid);
@@ -89,7 +89,7 @@ async function router(schema, config) {
         res: 'res.StackQueue.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             const iter = await Iteration.from(config.pool, req.params.iterationid);
