+++ api/routes/project-iteration-stack.js
@@ -2,7 +2,7 @@
 const { Err } = require('@openaddresses/batch-schema');
 const Iteration = require('../lib/project/iteration');
 const Stack = require('../lib/stack');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
 
@@ -24,7 +24,7 @@ async function router(schema, config) {
         res: 'res.Stack.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             const stack = await Stack.from(req.params.pid, req.params.iterationid);
@@ -54,7 +54,7 @@ async function router(schema, config) {
         res: 'res.Stack.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             const iter = await Iteration.from(config.pool, req.params.iterationid);
@@ -102,7 +102,7 @@ async function router(schema, config) {
         res: 'res.Standard.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
             config.is_aws();
 
             const stack = await Stack.from(req.params.pid, req.params.iterationid);
