+++ api/routes/project-iteration-submission.js
@@ -1,7 +1,7 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
 const Submission = require('../lib/project/iteration/submission');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
     const TileBase = (await import('tilebase')).default;
@@ -26,7 +26,7 @@ async function router(schema, config) {
         res: 'res.ListSubmission.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             res.json(await Submission.list(config.pool, req.params.iterationid, req.query));
         } catch (err) {
@@ -53,7 +53,7 @@ async function router(schema, config) {
         res: 'res.Task.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const sub = await Submission.from(config.pool, req.params.subid, req.params.pid);
 
@@ -82,7 +82,7 @@ async function router(schema, config) {
         res: 'res.TileJSON.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const sub = await Submission.from(config.pool, req.params.subid, req.params.pid);
 
@@ -124,7 +124,7 @@ async function router(schema, config) {
         try {
             req.user = req.token;
 
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             if (!['mvt', 'png'].includes(req.params.format)) {
                 throw new Err(400, null, '.mvt and .png are supported');
