+++ api/routes/project-iteration-task.js
@@ -1,7 +1,7 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
 const Task = require('../lib/project/iteration/task');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
 
@@ -25,7 +25,7 @@ async function router(schema, config) {
         res: 'res.Task.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             let task;
             if (req.body.type === 'vectorize') {
@@ -71,7 +71,7 @@ async function router(schema, config) {
         res: 'res.ListTasks.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             res.json(await Task.list(config.pool, req.params.iterationid, req.query));
         } catch (err) {
@@ -98,7 +98,7 @@ async function router(schema, config) {
         res: 'res.Task.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const task = await Task.from(config.pool, req.params.taskid);
 
@@ -127,7 +127,7 @@ async function router(schema, config) {
         res: 'res.Task.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const task = await Task.from(config.pool, req.params.taskid);
 
@@ -158,7 +158,7 @@ async function router(schema, config) {
         res: 'res.Task.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const task = await Task.from(config.pool, req.params.taskid);
             task.patch(req.body);
@@ -189,7 +189,7 @@ async function router(schema, config) {
         res: 'res.Standard.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const task = await Task.from(config.pool, req.params.taskid);
             await task.delete(config.pool);
