+++ api/routes/stacks.js
@@ -1,7 +1,7 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
 const Stack = require('../lib/stack');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
 
@@ -23,7 +23,7 @@ async function router(schema, config) {
         res: 'res.ListStacks.json'
     }, async (req, res) => {
         try {
-            await User.is_admin(req);
+            await Auth.is_admin(req);
 
             if (config.Environment !== 'aws') {
                 return {
