+++ api/routes/tokens.js
@@ -1,7 +1,7 @@
 'use strict';
 const { Err } = require('@openaddresses/batch-schema');
 const UserToken = new require('../lib/token');
-const User = require('../lib/user');
+const Auth = require('../lib/auth');
 
 async function router(schema, config) {
 
@@ -23,7 +23,7 @@ async function router(schema, config) {
         res: 'res.ListTokens.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             req.query.uid = req.user.id;
             return res.json(await UserToken.list(config.pool, req.query));
@@ -50,7 +50,7 @@ async function router(schema, config) {
         res: 'res.CreateToken.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             req.body.uid = req.user.id;
             return res.json((await UserToken.generate(config.pool, req.body)).serialize(true));
@@ -76,7 +76,7 @@ async function router(schema, config) {
         res: 'res.Token.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             let token = await UserToken.from(config.pool, req.params.token_id);
             if (token.uid !== req.user.id) throw new Err(401, null, 'Cannot get a token you did not create');
@@ -106,7 +106,7 @@ async function router(schema, config) {
         res: 'res.Standard.json'
     }, async (req, res) => {
         try {
-            await User.is_auth(req);
+            await Auth.is_auth(req);
 
             const token = await UserToken.from(config.pool, req.params.token_id);
             if (token.uid !== req.user.id) throw new Err(401, null, 'Cannot delete a token you did not create');
